Substring(s) of an ANSI colored string

This is the color-aware counterpart of base::substring(). It works exactly like the original, but keeps the colors in the substrings. The ANSI escape sequences are ignored when calculating the positions within the string.

col_substring(text, first, last = 1000000L)

Arguments

text

Character vector, potentially ANSI styled, or a vector to coarced to character. It is recycled to the longest of first and last.

first

Starting index or indices, recycled to match the length of x.

last

Ending index or indices, recycled to match the length of x.

Value

Character vector of the same length as x, containing the requested substrings. ANSI styles are retained.

See also

Examples

str <- paste(
  red("red"),
  "default",
  green("green")
)

cat(str, "\n")
#> red default green 
cat(col_substring(str, 1, 5), "\n")
#> red d 
cat(col_substring(str, 1, 15), "\n")
#> red default gre 
cat(col_substring(str, 3, 7), "\n")
#> d def 

substring(strip_style(str), 1, 5)
#> [1] "red d"
substring(strip_style(str), 1, 15)
#> [1] "red default gre"
substring(strip_style(str), 3, 7)
#> [1] "d def"

str2 <- "another " %+%
  red("multi-", sep = "", underline("style")) %+%
  " text"

cat(str2, "\n")
#> another multi-style text 
cat(col_substring(str2, c(3,5), c(7, 18)), sep = "\n")
#> other
#> her multi-styl
substring(strip_style(str2), c(3,5), c(7, 18))
#> [1] "other"          "her multi-styl"