show_ansi_colors {crayon}	R Documentation

Show the ANSI color table on the screen

Description

Show the ANSI color table on the screen

Usage

show_ansi_colors(colors = num_colors())

Arguments

colors
Number of colors to show, meaningful values are 8 and 256. It is automatically set to the number of supported colors, if not specified.

Value

The printed string, invisibly.

[Package crayon version 1.4.2 Index]