style {crayon}	R Documentation

Add style to a string

Description

See names(styles), or the crayon manual for available styles.

Usage

style(string, as = NULL, bg = NULL)

Arguments

string
Character vector to style.

as
Style function to apply, either the function object, or its name, or an object to pass to make_style().

bg
Background style, a style function, or a name that is passed to make_style().

Value

Styled character vector.

Examples

## These are equivalent
style("foobar", bold)
style("foobar", "bold")
bold("foobar")

[Package crayon version 1.4.2 Index]