styles {crayon}	R Documentation

ANSI escape sequences of crayon styles

Description

You can use this function to list all availables crayon styles, via names(styles()), or to explicitly apply an ANSI escape seauence to a string.

Usage

styles()

Value

A named list. Each list element is a list of two strings, named ‘open’ and ‘close’.

See Also

crayon() for the beginning of the crayon manual.

Examples

names(styles())
cat(styles()[["bold"]]$close)

[Package crayon version 1.4.2 Index]