Count number of characters in an ANSI colored string

This is a color-aware counterpart of base::nchar(), which does not do well, since it also counts the ANSI control characters.

Usage

col_nchar(x, ...)

Arguments

x

Character vector, potentially ANSO styled, or a vector to be coarced to character.

...

Additional arguments, passed on to base::nchar() after removing ANSI escape sequences.

Value

Numeric vector, the length of the strings in the character vector.

See also

Examples

str <- paste(
  red("red"),
  "default",
  green("green")
)

cat(str, "\n")
#> red default green 
nchar(str)
#> [1] 37
col_nchar(str)
#> [1] 17
nchar(strip_style(str))
#> [1] 17