concat {crayon}	R Documentation

Concatenate character vectors

Description

The length of the two arguments must match, or one of them must be of length one. If the length of one argument is one, then the output's length will match the length of the other argument. See examples below.

Usage

lhs %+% rhs

Arguments

lhs
Left hand side, character vector.

rhs
Right hand side, character vector.

Value

Concatenated vectors.

Examples

"foo" %+% "bar"

letters[1:10] %+% chr(1:10)

letters[1:10] %+% "-" %+% chr(1:10)

## This is empty (unlike for parse)
character() %+% "*"

[Package crayon version 1.4.2 Index]