Concatenate character vectors

The length of the two arguments must match, or one of them must be of length one. If the length of one argument is one, then the output's length will match the length of the other argument. See examples below.

Usage

lhs %+% rhs

Arguments

lhs

Left hand side, character vector.

rhs

Right hand side, character vector.

Value

Concatenated vectors.

Examples

"foo" %+% "bar"
#> [1] "foobar"

letters[1:10] %+% chr(1:10)
#>  [1] "a1"  "b2"  "c3"  "d4"  "e5"  "f6"  "g7"  "h8"  "i9"  "j10"

letters[1:10] %+% "-" %+% chr(1:10)
#>  [1] "a-1"  "b-2"  "c-3"  "d-4"  "e-5"  "f-6"  "g-7"  "h-8"  "i-9"  "j-10"

## This is empty (unlike for parse)
character() %+% "*"
#> character(0)