Switch on or off a style

Make a style active. The text printed to the screen from now on will use this style.

Usage

# S3 method for class 'crayon'
start(x, ...)

finish(x, ...)

# S3 method for class 'crayon'
finish(x, ...)

Arguments

x

Style.

...

Ignored.

Details

This function is very rarely needed, e.g. for colored user input. For other reasons, just call the style as a function on the string.

Examples

## The input is red (if color is supported)
get_name <- function() {
  cat("Enter your name:", start(red))
  input <- readline()
  cat(finish(red))
  input
}
name <- get_name()
#> Enter your name: 
#> 
name
#> [1] ""